Add endpoints and messages for Staking service
Note that most methods simply return a
Transaction, since most of the methods will
modify storage. In this case, most of the methods
will need to go through a contract execution,
warranting the Transaction return type.

Additionally, Golang should not be concerned
with any of the information contained in
these transactions, so there are no
specific transaction messages being
returned, only the generic one.